suggested changes: move hook and added glsl literal trigger
