+++ src/experiments/77.instanced-grass/shaders.ts
@@ -1,4 +1,15 @@
-export const grassVertexShader = `
+import SimplexNoise from 'simplex-noise'
+
+const simplex = new SimplexNoise(Math.random)
+
+export function getYPosition(x: number, z: number) {
+  let y = 2 * simplex.noise2D(x / 50, z / 50)
+  y += 4 * simplex.noise2D(x / 100, z / 100)
+  y += 0.2 * simplex.noise2D(x / 10, z / 10)
+  return y
+}
+
+export const grassVertexShader = /* glsl */ `
 precision highp float;
 
 attribute vec3 position;
@@ -140,7 +151,7 @@ void main() {
 }
 `
 
-export const grassFragmentShader = `
+export const grassFragmentShader = /* glsl */ `
 precision highp float;
 
 uniform sampler2D map;
